Add: a boolean flag to switch on/off clustering
At the moment the user has to use the -select argument to prevent the
background clustering (even if they want the complete UML) and they
don't have a way to activate clustering if they are selecting the
elements they want to show in the UML.

This flag lets the user swtich on/off the clustering background in all
cases. It is on by default (which is the current default behavior for
complete UML generation).